requirements.txt
# Math Helper 프로덕션 의존성
# Python 3.8 이상 필요

# 웹 프레임워크
streamlit>=1.28.0

# 데이터 처리 (향후 확장용)
pandas>=2.0.0

# 수치 계산
numpy>=1.24.0

# 시각화
matplotlib>=3.7.0
plotly>=5.17.0

# 설정 파일 파싱
pyyaml>=6.0